Introcue server response formatting
Server response is formatted in a ruby way rather than ready for
javascript. The formatter takes care of that. Furthermore the business
logic that in the frontend a patient has only on address is taken care
of only in the formatter. The rest of the application is unaware that on
the server side a patient might have more than one address.