Add version information for webpacker
There was an issue with the stylesheet tag not being present when using
webpacker without a specific version (latest at that time: 4.0.2).

A threadpost (https://github.com/rails/webpacker/issues/1659#issuecomment-450974230)
mentioned that locking webpacker to a specific version did the trick.
I tried the same and it DID work.